Initial async work is looking good
I now use a thread and job queuing to compile in parallel. Each update
will not only schedule a new compilation job, but also check if there is
a finished job. If there is, the highlighting will be updated. This
allows me to highlight translation units with, I've tested, thousands of
lines of code on my little Macbook Air with no problem.