Added initial cursor following
This will only highlight up to 100 lines around the current cursor,
making it possible to work in very large files. Compilation will still
be slow, of course. As of now, it refreshes the highlighting whenever
the cursor moves, which is not ideal. It should only happen when
approaching a highlighting boundry.