Add test for config file loading
Add basic test harness and the first tests.
A custom target in cmake was chosen over ctest because ctest does not
print out test ouput by default and is in genral more trouble than it is
worth.
The tests check if the default config is returned in case there is no
file, and if the file itself is properly parsed.