Simplify finding lua
There are cmake find modules for lua 5.0 and 5.1. The current cmake
master however does not contain a FindLua52. Unless a distribution adds
this script in their repos, searching for lua 5.2 will always fail and
we will fall back to lua 5.1 if it is available. If it is not, the
generic FindLua.cmake will find lua 5.2. This is a problem on systems
where both 5.1 and 5.2 are installed, as 5.2 will never be selected. As
the FindLua.cmake searches for versions >= 5.0 and returns the highest
version it finds we can just use that.

The FindLua package only got added with CMake 3.0.0 so we have to add
our own.