Take fs::path as an argument instead of a string
We create temporary path objects almost every time filename is used.
This indicates that we may want to operate on a path in the first place.
The only time we actually need a string is when replacing the filename
at the end. The conversion to string is implicit in this case.