Only disable CXX11_ABI on downloaded clang
We know for a fact that our "trusted clang" is not build against the
new ABI but we cannot assume that this is the case for the system clang
install. In all likelihood the system install ist build against the same
ABI that the compiler uses so we should not need to change anything.