Add support for llvm 4.0
LLVM 4 has changed the function signature of loadFromFile to include a
syntax parameter that dictates how the command line of the compilation
database commands should be parsed. Seeing how the autodetection seems
to be flaky as indicated by the commit that introduced the change, we
should set the syntax to "Gnu" instead of "autodetect" to avoid problems
with our supported platforms.
This does of course mean that we can not support JSON compilation
databases on windows using llvm 4 or higher.