Fix compilation database with relative paths
Fixes #162
Loading a file that has a relative path in the compile arguments will
fail with an internal error because we only try to find and remove the
absolute path of the file from the compilation arguments.

Now we also try to remove the relative filename if both "file" and the
"command" contain paths relative to the "directory".

Note that this will fail in the case that the "file" is an absolute path
and the "argument" is still relative. For this to work we would have to
make the filename itself a relative path using something like
boost::filesystem::relative, which only got introduced in boost 1.6.